Update README.md
The user interacts with the tool mainly via Excel.  A spreadsheet is prepared with a list of PI Point names in one column of a worksheet and a list of event start times in another worksheet (ideally these times are instances of faults or anomalies).  The user enters this filepath in a text box on the form, selects the appropriate PI Data Archive, and chooses offset parameters that are used to construct a time range prior to each event time.  The tool then calculates the Coefficient of Variation for each PI tag and identifies the tags with the top 10 COV for each time range.  A summary is provided via output Excel workbook that ranks the PI Points based on how many times/events for which each tag was identified as having top 10 COV.